requirements.txt
discord.py
python-dotenv
google-genai
watchdog
colorlog
flask
requests
yt-dlp
httpx>=0.28.0
click
aiohttp
deep-translator